requirements.txt
streamlit
pymupdf
pytesseract
pillow
pymongo